+++ src/plc.rs
@@ -1,50 +1,37 @@
 use ads_client as ads;
 use bon::bon;
 use crossbeam_channel::Receiver;
-use std::{collections::HashMap, time::Duration};
+use std::{sync::Arc, time::Duration};
+use tokio::sync::Mutex;
 
-pub use ads::AdsTimeout;
-
-use crate::{
-    plc_connection::PlcConnection, PlcDataType, PlcParams, SymbolTypeTree, SymbolTypeTreeError,
-};
-
-#[derive(Debug, thiserror::Error)]
-pub enum PlcError {
-    #[error("ADS error {0}")]
-    Ads(#[from] ads::AdsError),
-
-    #[error("Symbol Type Tree error {0}")]
-    SymbolTypeTree(#[from] SymbolTypeTreeError),
-}
-
-#[derive(Debug, Copy, Clone)]
-pub struct SymbolHandle(u32);
-
-impl SymbolHandle {
-    pub fn as_u32(&self) -> u32 {
-        self.0
-    }
-}
-
-#[derive(Debug, Copy, Clone)]
-pub struct NotificationHandle(u32);
-
-impl NotificationHandle {
-    pub fn as_u32(&self) -> u32 {
-        self.0
-    }
-}
+use crate::{PlcConnection, PlcDataType, PlcError, PlcParams, SymbolTypeTree, SymbolTypeTreeError};
 
 pub struct Plc {
     addr: String,
     port: u16,
-    timeout: Option<AdsTimeout>,
+    timeout: Option<Timeout>,
     retry_delay: Option<Duration>,
     set_to_run_mode: bool,
-    connection: PlcConnection,
-    symbol_handles: HashMap<String, SymbolHandle>,
-    notification_handles: Vec<NotificationHandle>,
+    connection: Arc<Mutex<PlcConnection>>,
+}
+
+// Note(mw): We need our own Timeout enum, only because
+//   ads_client::AdsTimeout doesn't implement Clone.
+#[derive(Debug, Clone)]
+pub enum Timeout {
+    /// Default timeout is 5 seconds
+    DefaultTimeout,
+    /// Custom timeout (in seconds)
+    CustomTimeout(u64),
+}
+
+impl From<Timeout> for ads::AdsTimeout {
+    fn from(value: Timeout) -> Self {
+        match value {
+            Timeout::DefaultTimeout => ads::AdsTimeout::DefaultTimeout,
+            Timeout::CustomTimeout(timeout) => ads::AdsTimeout::CustomTimeout(timeout),
+        }
+    }
 }
 
 /// Provides a more user friendly wrapper to interact with the OpenStar PLC's.
@@ -54,7 +41,7 @@ impl Plc {
     pub fn new(
         addr: String,
         port: u16,
-        timeout: Option<AdsTimeout>,
+        timeout: Option<Timeout>,
         retry_delay: Option<Duration>,
         set_to_run_mode: bool,
     ) -> Self {
@@ -64,221 +51,385 @@ impl Plc {
             timeout,
             retry_delay,
             set_to_run_mode,
-            ads_client: Default::default(),
-            symbol_handles: Default::default(),
-            notification_handles: Default::default(),
+            connection: Default::default(),
         }
     }
 
-    async fn ads_client(&self) -> Result<&ads::Client, PlcError> {
-        let Some(client) = self.ads_client else {
-            let mut builder = ads::ClientBuilder::new(&self.addr, self.port);
-            if let Some(timeout) = self.timeout {
-                builder = builder.set_timeout(timeout);
-            };
-            if let Some(retry_delay) = self.retry_delay {
-                builder = builder.set_retry_delay(Some(retry_delay));
-            };
-            let client = builder.build().await?;
-            self.ads_client = Some(client);
-            return Ok(&client);
-        };
-        Ok(&client)
-    }
-
-    fn symbol_handle(&mut self, symbol: &str) -> Result<SymbolHandle, PlcError> {
-        // TODO: might need to think a bit more about other cases we may need to invalidate these handles e.g: new code flashed onto the PLC
-        let handle = self.symbol_handles.get(symbol);
-        let handle = match handle {
-            Some(handle) => handle.to_owned(),
-            None => {
-                let handle_raw = ads::Handle::new(self.ads_device(), symbol)?.raw();
-                let handle = SymbolHandle(handle_raw);
-                self.symbol_handles.insert(symbol.to_owned(), handle);
-                handle
+    /// Loops to keep a PLC connected over ADS.
+    pub async fn connection_loop(&self) {
+        loop {
+            if !self.is_connected() {
+                let mut connection = self.connection.lock().await;
+                if let Err(error) = connection
+                    .connect(
+                        &self.addr,
+                        self.port,
+                        self.timeout.clone().map(Into::into),
+                        self.retry_delay,
+                        self.set_to_run_mode,
+                    )
+                    .await
+                {
+                    println!("PLC connection failed, {}. Retrying in 2 seconds...", error);
+                } else {
+                    println!("PLC connection successful!");
+                }
             }
-        };
 
-        Ok(handle)
+            std::thread::sleep(Duration::from_secs(2));
+        }
     }
 
-    /// Returns if the connected ADS device is in run mode.
-    pub fn is_run_mode(&self) -> Result<bool, PlcError> {
-        let (state, _) = self.ads_device().get_state()?;
-
-        Ok(state == ads::AdsState::Run)
+    /// Disconnects the internal PLC client
+    pub async fn disconnect(&self) {
+        let mut connection = self.connection.lock().await;
+        connection.disconnect();
     }
 
-    /// Attempts to set the ADS device into run mode if it is not already in it.
-    pub fn set_to_run_mode(&self) -> Result<(), PlcError> {
-        let device = self.ads_device();
+    pub fn is_connected(&self) -> bool {
+        let connection = self.connection.lock().await;
+        match *connection {
+            PlcConnection::Connected(_) => true,
+            PlcConnection::Disconnected => false,
+        }
+    }
 
-        println!("Device Info: {:?}", device.get_info());
+    /// Gets a symbol type tree for a given symbol path.
+    ///
+    /// Returns None if the PLC is not connected.
+    /// Returns any errors from the PLC
+    pub async fn get_dynamic_type_tree(&self, name: &str) -> Result<Option<SymbolTypeTree>> {
+        let mut connection = self.connection.lock().await;
 
-        let (state, dev_state) = device.get_state()?;
+        if let Some(client) = connection.client_mut() {
+            let type_tree = client.get_dynamic_type_tree(name).map_err(|error| {
+                println!(
+                    "PLC client error when getting information for symbol {}: {}",
+                    name, error
+                );
 
-        println!("Device State: {:?}", (state, dev_state));
+                match &error {
+                    PlcError::SymbolTypeTree(_symbol_type_tree_error) => connection.disconnect(),
+                    error => connection.handle_disconnect_error(error),
+                };
 
-        if state != ads::AdsState::Run {
-            println!("Attempting to set PLC to run mode...");
+                error
+            })?;
 
-            device.write_control(ads::AdsState::Run, dev_state)?;
-
-            println!("Device State: {:?}", device.get_state());
+            return Ok(Some(type_tree));
         }
 
-        assert!(device.get_state()?.0 == ads::AdsState::Run);
-
-        Ok(())
+        Ok(None)
     }
 
-    /// Read the value of a symbol with a given type once.
-    pub fn read_symbol<T: PlcDataType>(&self, symbol: &str) -> Result<T, PlcError> {
-        let mut read_data = T::default();
+    /// Subscribes to the given symbol using the symbol type tree and sends deserialised tree-like data back with the sender channel.
+    ///
+    /// Returns None if the PLC is not connected.
+    /// Returns any errors from the PLC.
+    pub async fn start_dynamic_symbol_receiver(
+        &self,
+        name: String,
+        symbol_type_tree: SymbolTypeTree,
+        sender_channel: broadcast::Sender<SymbolTree>,
+    ) -> Result<Option<()>> {
+        let mut notif_handle = None;
+        {
+            let mut connection = self.connection.lock().await;
 
-        let index_offset = self.symbol_handle(symbol)?.as_u32();
+            if let Some(client) = connection.client_mut() {
+                notif_handle = Some(
+                    client
+                        .add_dynamic_symbol_notification(&name, &symbol_type_tree)
+                        .map_err(|error| {
+                            println!(
+                                "PLC client error when getting information for symbol {}: {}",
+                                name, error
+                            );
 
-        self.ads_device().read_exact(
-            ads::index::RW_SYMVAL_BYHANDLE,
-            index_offset,
-            read_data.as_bytes_mut(),
-        )?;
+                            connection.handle_disconnect_error(&error);
 
-        Ok(read_data)
+                            error
+                        })?,
+                );
+            }
+        }
+
+        let Some(mut notif_receiver) = self.notification_receiver() else {
+            return Ok(None);
+        };
+
+        let Some(notif_handle) = notif_handle else {
+            return Ok(None);
+        };
+
+        // Runs the actual notification loop which filters for the correct symbol handle.
+        loop {
+            let resp =
+                tokio::task::spawn_blocking(move || Self::recv_blocking(notif_receiver)).await.context("Tokio spawn blocking function failed while trying to receive from PLC channel.")?;
+            notif_receiver = resp.0;
+            let notif_result = resp.1;
+            let notif = notif_result
+                .context("Receive error while receiving from PLC notification channel.")?;
+            for sample in notif.samples() {
+                if sample.handle == notif_handle {
+                    let symbol_tree: SymbolTree = (&symbol_type_tree, sample.data, 0).into();
+                    // If there is an error sending it means that all receivers are gone and therefore this thread has successfully ended.
+                    if let Err(_err) = sender_channel.send(symbol_tree) {
+                        return Ok(Some(()));
+                    };
+                }
+            }
+        }
     }
 
-    /// Invokes a PLC method (which has the attribute 'TcRpcEnable') with parameters.
-    pub fn invoke_rpc_method<Params: PlcParams>(
-        &mut self,
-        symbol: &str,
-        params: Params,
-    ) -> Result<(), PlcError> {
-        let index_offset = self.symbol_handle(symbol)?.as_u32();
-        let write_data = params.as_data();
+    /// Subscribes to the given symbol using the symbol type tree and sends deserialised flattened map data back with the sender channel.
+    /// The key to the map is the path of the symbol relative to the named symbol provided.
+    ///
+    /// Returns None if the PLC is not connected.
+    /// Returns any errors from the PLC.
+    pub async fn start_dynamic_symbol_map_receiver(
+        &self,
+        name: String,
+        symbol_type_tree: SymbolTypeTree,
+        sender_channel: broadcast::Sender<SymbolMap>,
+    ) -> Result<Option<()>> {
+        let mut notif_handle = None;
+        {
+            let mut connection = self.connection.lock().await;
 
-        self.ads_device().write_read_exact(
-            ads::index::RW_SYMVAL_BYHANDLE,
-            index_offset,
-            &write_data,
-            &mut [],
-        )?;
+            if let Some(client) = connection.client_mut() {
+                notif_handle = Some(
+                    client
+                        .add_dynamic_symbol_notification(&name, &symbol_type_tree)
+                        .map_err(|error| {
+                            println!(
+                                "PLC client error when getting information for symbol {}: {}",
+                                name, error
+                            );
 
-        Ok(())
+                            connection.handle_disconnect_error(&error);
+
+                            error
+                        })?,
+                );
+            }
+        }
+
+        let Some(mut notif_receiver) = self.notification_receiver() else {
+            return Ok(None);
+        };
+
+        let Some(notif_handle) = notif_handle else {
+            return Ok(None);
+        };
+
+        let symbol_type_map = SymbolTypeMap::from_tree(symbol_type_tree, 0, "".to_string());
+        // Runs the actual notification loop which filters for the correct symbol handle.
+        loop {
+            let resp =
+                tokio::task::spawn_blocking(move || Self::recv_blocking(notif_receiver)).await.context("Tokio spawn blocking function failed while trying to receive from PLC channel.")?;
+            notif_receiver = resp.0;
+            let notif_result = resp.1;
+            let notif = notif_result
+                .context("Receive error while receiving from PLC notification channel.")?;
+            for sample in notif.samples() {
+                if sample.handle == notif_handle {
+                    let symbol_tree: SymbolMap =
+                        SymbolMap::from_bytes(&symbol_type_map, sample.data);
+                    // If there is an error sending it means that all receivers are gone and therefore this thread has successfully ended.
+                    if let Err(_err) = sender_channel.send(symbol_tree) {
+                        return Ok(Some(()));
+                    };
+                }
+            }
+        }
     }
 
-    /// Invokes a PLC method (which has the attribute 'TcRpcEnable') and returns the resulting value.
-    pub fn fetch_from_rpc_method<Params: PlcParams, Value: PlcDataType>(
-        &mut self,
-        symbol: &str,
-        params: Params,
-    ) -> Result<Value, PlcError> {
-        let index_offset = self.symbol_handle(symbol)?.as_u32();
-        let write_data = params.as_data();
-        let mut read_data = Value::default();
-
-        self.ads_device().write_read_exact(
-            ads::index::RW_SYMVAL_BYHANDLE,
-            index_offset,
-            &write_data,
-            read_data.as_bytes_mut(),
-        )?;
-
-        Ok(read_data)
+    // Fully takes ownership of the receiver so that this can be run within a spawn_blocking enclosure.
+    // Returns out the channel again with the result so that the caller can ensure the channel can be used again by awaiting the response.
+    fn recv_blocking<T>(channel: Receiver<T>) -> (Receiver<T>, Result<T>) {
+        let result = channel.recv_timeout(std::time::Duration::from_secs(5));
+        let value = match result {
+            Ok(value) => anyhow::Result::Ok(value),
+            Err(err) => match err {
+                crossbeam_channel::RecvTimeoutError::Timeout => {
+                    anyhow::Result::Err(anyhow!("Message took more than 5 seconds to receive."))
+                }
+                crossbeam_channel::RecvTimeoutError::Disconnected => {
+                    anyhow::Result::Err(anyhow!("PLC channel closed."))
+                }
+            },
+        };
+        (channel, value)
     }
 
-    /// Subscribes to a symbol and returns the notification handle.
-    pub fn subscribe<T: PlcDataType>(
-        &mut self,
+    /// Read a symbol from the PLC.
+    ///
+    /// Returns None if the PLC is not connected.
+    pub async fn read_symbol<T: PlcDataType>(&self, name: &str) -> Result<Option<T>> {
+        let mut connection = self.connection.lock().await;
+
+        if let Some(client) = connection.client_mut() {
+            let value = client.read_symbol(name).map_err(|error| {
+                println!("PLC client error when reading symbol {}: {}", name, error);
+
+                connection.handle_disconnect_error(&error);
+
+                error
+            })?;
+
+            return Ok(Some(value));
+        }
+
+        Ok(None)
+    }
+
+    /// Calls an RPC method on the PLC that returns a value.
+    ///
+    /// Returns None if the PLC is not connected.
+    pub async fn fetch_from_rpc_method<T: PlcDataType>(&self, name: &str) -> Result<Option<T>> {
+        let mut connection = self.connection.lock().await;
+
+        if let Some(client) = connection.client_mut() {
+            let value = client.fetch_from_rpc_method(name).map_err(|error| {
+                eprintln!(
+                    "PLC client error when invoking RPC method {}: {}",
+                    name, error
+                );
+
+                connection.handle_disconnect_error(&error);
+
+                error
+            })?;
+
+            return Ok(Some(value));
+        }
+
+        Ok(None)
+    }
+
+    /// Calls an RPC method on the PLC.
+    ///
+    /// Returns None if the PLC is not connected.
+    pub async fn invoke_rpc_method(&self, name: &str) -> Result<Option<()>> {
+        let mut connection = self.connection.lock().await;
+
+        if let Some(client) = connection.client_mut() {
+            client.invoke_rpc_method(name).map_err(|error| {
+                eprintln!(
+                    "PLC client error when invoking RPC method {}: {}",
+                    name, error
+                );
+
+                connection.handle_disconnect_error(&error);
+
+                error
+            })?;
+
+            return Ok(Some(()));
+        }
+
+        Ok(None)
+    }
+
+    /// Calls an RPC method on the PLC with one parameter.
+    ///
+    /// Returns None if the PLC is not connected.
+    pub fn invoke_rpc_method_with_param<P: PlcDataType>(
+        &self,
         name: &str,
-    ) -> Result<NotificationHandle, PlcError> {
-        let notification_handle_raw = {
-            let index_offset = self.symbol_handle(name)?.as_u32();
+        param: P,
+    ) -> Result<Option<()>> {
+        let mut connection = self.connection.lock().await;
 
-            self.ads_device().add_notification(
-                ads::index::RW_SYMVAL_BYHANDLE,
-                index_offset,
-                &ads::notif::Attributes::new(
-                    T::size(),
-                    ads::notif::TransmissionMode::ServerOnChange,
-                    std::time::Duration::ZERO,
-                    // TODO: setting this to higher e.g: 1000ms does not work, maybe because the status data is changing every PLC cycle?
-                    // NB: Setting this to 10ms to match the PLC cycle time that it seems to be reporting at anyway
-                    std::time::Duration::from_millis(10),
-                ),
-            )?
-        };
-        let notification_handle = NotificationHandle(notification_handle_raw);
-        self.notification_handles.push(notification_handle);
+        if let Some(client) = connection.client_mut() {
+            client
+                .invoke_rpc_method_with_param(name, param)
+                .map_err(|error| {
+                    eprintln!(
+                        "PLC client error when invoking RPC method {}: {}",
+                        name, error
+                    );
 
-        Ok(notification_handle)
-    }
+                    connection.handle_disconnect_error(&error);
 
-    /// Gets a type tree for the symbol to get the format of a symbol's internal structure at runtime.
-    pub fn get_dynamic_type_tree(&self, name: &str) -> Result<SymbolTypeTree, PlcError> {
-        let (symbols, type_map) = ads::symbol::get_symbol_info(self.ads_device())?;
-        let mut symbol = None;
+                    error
+                })?;
 
-        for symbol_info in symbols {
-            if symbol_info.name.to_lowercase() == name.to_lowercase() {
-                symbol = Some(symbol_info);
-            }
+            return Ok(Some(()));
         }
 
-        let Some(symbol) = symbol else {
-            return Ok(SymbolTypeTree::Missing);
-        };
-
-        let symbol_type_tree = match (&symbol, &type_map).try_into() {
-            Ok(symbol_type_tree) => symbol_type_tree,
-            Err(err) => {
-                println!("Error when getting symbol type from type map {err:?}");
-                return Ok(SymbolTypeTree::Unknown(symbol.size));
-            }
-        };
-
-        Ok(symbol_type_tree)
+        Ok(None)
     }
 
-    /// Subscribes to a symbol based off of its type tree and returns the handle as a u32.
-    pub fn add_dynamic_symbol_notification(
-        &mut self,
-        symbol: &str,
-        symbol_type_tree: &SymbolTypeTree,
-    ) -> Result<u32, PlcError> {
-        let index_offset = self.symbol_handle(symbol)?.as_u32();
+    /// Calls an RPC method on the PLC with three parameters.
+    ///
+    /// Returns None if the PLC is not connected.
+    pub fn invoke_rpc_method_with_three_params<
+        P1: PlcDataType,
+        P2: PlcDataType,
+        P3: PlcDataType,
+    >(
+        &self,
+        name: &str,
+        param_1: P1,
+        param_2: P2,
+        param_3: P3,
+    ) -> Result<Option<()>> {
+        let mut connection = self.connection.lock().await;
 
-        let notif_handle = self.ads_device().add_notification(
-            ads::index::RW_SYMVAL_BYHANDLE,
-            index_offset,
-            &ads::notif::Attributes::new(
-                symbol_type_tree.get_size(),
-                ads::notif::TransmissionMode::ServerOnChange,
-                std::time::Duration::ZERO,
-                std::time::Duration::from_millis(10),
-            ),
-        )?;
+        if let Some(client) = connection.client_mut() {
+            client
+                .invoke_rpc_method_with_three_params(name, param_1, param_2, param_3)
+                .map_err(|error| {
+                    eprintln!(
+                        "PLC client error when invoking RPC method {}: {}",
+                        name, error
+                    );
 
-        Ok(notif_handle)
-    }
+                    connection.handle_disconnect_error(&error);
 
-    /// Get a crossbeam channel receiver for all ADS notifications.
-    pub fn notification_receiver(&self) -> Receiver<ads::notif::Notification> {
-        self.ads_client().get_notification_channel()
-    }
+                    error
+                })?;
 
-    /// Deletes an ongoing subscription using its handle.
-    pub fn unsubscribe(&self, notification_handle: u32) {
-        // NB: unsure why, but deleting the notification returns a "Notification handle is invalid" error, hence the ok() here.
-        // It still works though, so maybe not a problem.
-        self.device().delete_notification(notification_handle).ok();
-    }
-
-    /// Deletes all ongoing symbol subscriptions.
-    pub fn unsubscribe_all(&mut self) {
-        for notification_handle in &self.notification_handles {
-            self.unsubscribe(*notification_handle);
+            return Ok(Some(()));
         }
 
-        self.notification_handles.clear();
+        Ok(None)
+    }
+
+    /// Subscribes to a notification channel on the PLC, returning a handle to the channel.
+    ///
+    /// Returns None if the PLC is not connected.
+    pub fn subscribe<T: PlcDataType>(&self, name: &str) -> Result<Option<u32>> {
+        let mut connection = self.connection.lock().await;
+
+        if let Some(client) = connection.client_mut() {
+            let handle = client.subscribe::<T>(name).map_err(|error| {
+                eprintln!(
+                    "PLC client error when subscribing to notifications from {}: {}",
+                    name, error
+                );
+
+                connection.handle_disconnect_error(&error);
+
+                error
+            })?;
+
+            return Ok(Some(handle));
+        }
+
+        Ok(None)
+    }
+
+    /// Gets a notification receiver that streams symbol data as it changes on the PLC.
+    ///
+    /// A symbol must first be subscribed using the subscribe function.
+    pub fn notification_receiver(&self) -> Option<Receiver<ads::notif::Notification>> {
+        let connection = self.connection.lock().await;
+
+        connection
+            .client()
+            .map(|client| client.notification_receiver())
     }
 }
